+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-dev-master (2015-06-16)
-------------------
-* Initial development
+[dev-master] - 2015-10-14
+-------------------------
+
+### Changed
+- Removed 'Logger' classes
+- Added 'register' and 'unregister' methods
+
+[dev-master] - 2015-06-16
+-------------------------
+
+### Added
+- Initial development